container-data-model/app/requirements.txt
Flask==2.3.3
Flask-Swagger-UI==4.11.1
mysql-connector-python==8.0.28